Rearranged the game state list order.
Added an UpdateProgression function for map level.
Added fake bools for arm/hand unlock and clear states.
Removed warning in getting mouse position.
Completely updated the map level with current progression and fade overlays.
Updated player save/load functions for use with fake bools and reduced the drop down time.
Increased the default Zindex on buttons.

+++ MansionMashers/MansionMashers/SourceFiles/Input.c
@@ -118,9 +118,9 @@ void FoxInput_GetWorldPosition(int *mx, int *my)
 	AEGfxGetCamPosition(&camX, &camY);
 
 	if (mx != NULL)
-		*mx = mouseX + camX - GetWindowWidth() / 2;
+		*mx = mouseX + (int)camX - GetWindowWidth() / 2;
 	if (my != NULL)
-		*my = mouseY + camY - GetWindowHeight() / 2;
+		*my = mouseY + (int)camY - GetWindowHeight() / 2;
 }
 
 int FoxInput_MouseWithinWindow(void)
